backend/requirements.txt
django
djangorestframework
djangorestframework_simplejwt
Faker